move References from instances.py to references.py
